[#27] 오디오 재생 관련 기능 AudioPlayViewModelable로 분리
오디오 재생 관련 중복 코드를 줄이기 위해 AudioPlayViewModelable 타입 구현
